requirements.txt
Pillow==9.2.0
pyocr==0.8.3
tesseract==0.1.3
